requirements.txt
streamlit
tensorflow
yfinance
pandas
numpy
matplotlib
scikit-learn
snowflake-connector-python[pandas]
